refactor: jpa repo 패키지 생성 및 분리
