+++ backend/src/main/java/solid/backend/jpaRepository/BasketRepository.java
@@ -1,0 +1,7 @@
+package solid.backend.jpaRepository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import solid.backend.entity.Basket;
+
+public interface BasketRepository extends JpaRepository<Basket, Integer> {
+}
