+++ backend/src/main/java/solid/backend/jpaRepository/PaymentRepository.java
@@ -1,0 +1,7 @@
+package solid.backend.jpaRepository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import solid.backend.entity.Payment;
+
+public interface PaymentRepository extends JpaRepository<Payment, Integer> {
+}
