+++ backend/src/main/java/solid/backend/jpaRepository/TravelRepository.java
@@ -1,4 +1,4 @@
-package solid.backend.admin.travel.repository;
+package solid.backend.jpaRepository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import solid.backend.entity.Travel;
